zam-repondeur-1.17.7/deploy_requirements.txt
-i https://pypi.org/simple
alembic==1.6.5
attrs==21.2.0
babel==2.9.1
beautifulsoup4==4.10.0
bleach==4.0.0
Brotli==1.0.9
CacheControl==0.12.6
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cssselect2==0.4.1
Cython==0.29.24
dataclasses==0.6
defusedxml==0.7.1
et-xmlfile==1.1.0
html5lib==1.1
huey==2.2.0
hupper==1.10.3
idna==3.2
inscriptis==2.0.0
Jinja2==3.0.1
htmlmin==0.1.12
lockfile==0.12.2
lxml==4.6.5
Mako==1.1.5
MarkupSafe==2.0.1
more-itertools==8.8.0
msgpack==1.0.2
openpyxl==3.0.7
parsy==1.3.0
PasteDeploy==2.1.1
pathvalidate==2.4.1
Pillow==9.0.1
pikepdf==5.1.3
plaster==1.0
plaster-pastedeploy==0.7
progressist==0.1.0
psycopg2-binary==2.9.1
pycparser==2.20
pydyf==0.1.1
pyphen==0.11.0
pyramid==1.10.8  # < 2.0 https://docs.pylonsproject.org/projects/pyramid/en/latest/whatsnew-2.0.html#upgrading-auth-20
# pyramid-debugtoolbar==4.9
pyramid-default-cors==0.2.1
pyramid-jinja2==2.8
pyramid-mailer==0.15.1
pyramid-retry==2.1.1
pyramid-tm==2.4
python-dateutil==2.8.2
python-editor==1.0.4
python-redis-lock==3.7.0
python-slugify==5.0.2
python-throttle==0.2.0
pytz==2019.3
redis==3.5.3
regex==2021.9.24
repoze.sendmail==4.4.1
requests==2.26.0
rollbar==0.14.7
selectolax==0.3.1
six==1.16.0
SQLAlchemy==1.3.24        # < 1.4 un passage en 1.4 induit des 'SAWarning: implicitly coercing SELECT object to scalar subquery'
SQLAlchemy-Utils==0.37.8
text-unidecode==1.3
tinycss2==1.1.0
transaction==3.0.1
translationstring==1.4
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.6
venusian==3.0.0
WeasyPrint==52.5
webencodings==0.5.1
WebOb==1.8.7
xmltodict==0.12.0
xvfbwrapper==0.2.9
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.5
zopfli==0.1.8
gunicorn==19.10.0